Reworked TesterAPI to support prog and patgen
