+++ example/targets/eagle_with_smt7.py
@@ -1,2 +1,2 @@
 origen.app.instantiate_dut("dut.eagle")
-origen.tester.target("::V93K::ST7")
+origen.tester.target("::V93K::SMT7")
