+++ example/targets/tester/uflex.py
@@ -1,0 +1,1 @@
+origen.tester.target("::ULTRAFLEX")
