+++ example/targets/tester/v93k_smt7.py
@@ -1,0 +1,1 @@
+origen.tester.target("::V93K::SMT7")
