+++ example/targets/tester/v93k_smt8.py
@@ -1,0 +1,1 @@
+origen.tester.target("::V93K::SMT8")
