--- example/targets/tester/v93k_st7.py
@@ -1,1 +1,0 @@
-origen.tester.target("::V93K::ST7")
