+++ rust/origen/src/testers/simulator/mod.rs
@@ -43,7 +43,8 @@ impl TesterAPI for Renderer {
         Box::new(std::clone::Clone::clone(self))
     }
 
-    fn run(&mut self, node: &Node) -> crate::Result<Node> {
-        Ok(node.process(self)?.unwrap())
+    fn render_pattern(&mut self, ast: &Node) -> crate::Result<()> {
+        ast.process(self)?;
+        Ok(())
     }
 }
